backend/services/calendar/requirements.txt
Flask
Flask-SQLAlchemy
PyJWT
requests
python-dotenv
cryptography
psycopg2-binary